requirements.txt
# requirements.txt
Flask
requests
fpdf
google-genai